feat: memo 안에 있던 코드를 memoApi, tagApi로 분리
